# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Challenge Arena API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type RegisterBodyGender =
  | (typeof RegisterBodyGender)[keyof typeof RegisterBodyGender]
  | null;

export const RegisterBodyGender = {
  male: "male",
  female: "female",
  other: "other",
} as const;

export interface RegisterBody {
  /**
   * @minLength 3
   * @maxLength 30
   */
  username: string;
  /** @minLength 6 */
  password: string;
  email?: string;
  freefireUid: string;
  ign?: string;
  gender?: RegisterBodyGender;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface PlayerStats {
  userId: string;
  matchesPlayed: number;
  wins: number;
  losses: number;
  winStreak: number;
  weeklyWins: number;
}

export interface User {
  id: string;
  username: string;
  freefireUid?: string | null;
  ign?: string | null;
  gender?: string | null;
  createdAt: string;
  email?: string | null;
  whatsappPhone?: string | null;
  /** True when user may access admin API (env or Firestore flag) */
  isAdmin?: boolean;
  stats?: PlayerStats | null;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type UpdateUserBodyGender =
  | (typeof UpdateUserBodyGender)[keyof typeof UpdateUserBodyGender]
  | null;

export const UpdateUserBodyGender = {
  male: "male",
  female: "female",
  other: "other",
} as const;

export interface UpdateUserBody {
  freefireUid?: string;
  ign?: string;
  gender?: UpdateUserBodyGender;
}

export type ChallengeMode = (typeof ChallengeMode)[keyof typeof ChallengeMode];

export const ChallengeMode = {
  "1v1": "1v1",
  "2v2": "2v2",
  "4v4": "4v4",
} as const;

export type ChallengeStatus =
  (typeof ChallengeStatus)[keyof typeof ChallengeStatus];

export const ChallengeStatus = {
  open: "open",
  full: "full",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
  disputed: "disputed",
} as const;

export interface TeamMember {
  userId: string;
  username: string;
  ign?: string | null;
  isLeader: boolean;
  joinedAt: string;
}

export interface Team {
  id: string;
  name: string;
  leaderId: string;
  players: TeamMember[];
  maxSize: number;
}

export interface Challenge {
  id: string;
  title: string;
  mode: ChallengeMode;
  scheduledAt: string;
  rules: string[];
  customRule?: string | null;
  status: ChallengeStatus;
  teamA: Team;
  teamB?: Team | null;
  roomId?: string | null;
  roomPassword?: string | null;
  createdAt: string;
  creatorId: string;
  winnerId?: string | null;
}

export type CreateChallengeBodyMode =
  (typeof CreateChallengeBodyMode)[keyof typeof CreateChallengeBodyMode];

export const CreateChallengeBodyMode = {
  "1v1": "1v1",
  "2v2": "2v2",
  "4v4": "4v4",
} as const;

export interface CreateChallengeBody {
  title: string;
  mode: CreateChallengeBodyMode;
  scheduledAt: string;
  rules: string[];
  customRule?: string | null;
  teamName: string;
}

export type JoinChallengeBodySide =
  (typeof JoinChallengeBodySide)[keyof typeof JoinChallengeBodySide];

export const JoinChallengeBodySide = {
  teamA: "teamA",
  teamB: "teamB",
} as const;

export interface JoinChallengeBody {
  side: JoinChallengeBodySide;
  /** Team name (required when joining as team B leader) */
  teamName?: string;
}

export type SubmitResultBodyWinningSide =
  (typeof SubmitResultBodyWinningSide)[keyof typeof SubmitResultBodyWinningSide];

export const SubmitResultBodyWinningSide = {
  teamA: "teamA",
  teamB: "teamB",
} as const;

export interface SubmitResultBody {
  winningSide: SubmitResultBodyWinningSide;
  screenshotUrl?: string | null;
}

export interface ShareRoomBody {
  roomId: string;
  roomPassword: string;
}

export type MatchResultStatus =
  (typeof MatchResultStatus)[keyof typeof MatchResultStatus];

export const MatchResultStatus = {
  pending: "pending",
  confirmed: "confirmed",
  disputed: "disputed",
} as const;

export interface MatchResult {
  id: string;
  challengeId: string;
  submittedBy: string;
  winningSide: string;
  screenshotUrl?: string | null;
  status: MatchResultStatus;
  createdAt: string;
}

export interface Message {
  id: string;
  challengeId: string;
  senderId: string;
  senderUsername: string;
  content: string;
  createdAt: string;
}

export interface SendMessageBody {
  /**
   * @minLength 1
   * @maxLength 500
   */
  content: string;
}

export type NotificationType =
  (typeof NotificationType)[keyof typeof NotificationType];

export const NotificationType = {
  player_joined: "player_joined",
  team_almost_full: "team_almost_full",
  match_ready: "match_ready",
  match_starting: "match_starting",
  match_result: "match_result",
} as const;

export interface Notification {
  id: string;
  userId: string;
  type: NotificationType;
  title: string;
  message: string;
  isRead: boolean;
  challengeId?: string | null;
  createdAt: string;
}

export interface LeaderboardEntry {
  rank: number;
  userId: string;
  username: string;
  ign?: string | null;
  wins: number;
  winStreak: number;
  matchesPlayed: number;
  badge?: string | null;
}

export interface AdminRecentUser {
  id: string;
  username: string;
  ign?: string | null;
  createdAt: string;
}

export type AdminOverviewChallengesByStatus = { [key: string]: number };

export interface AdminOverview {
  usersTotal: number;
  challengesTotal: number;
  challengesByStatus: AdminOverviewChallengesByStatus;
  teamsTotal: number;
  teamMembersTotal: number;
  messagesTotal: number;
  notificationsTotal: number;
  unreadNotifications: number;
  matchResultsTotal: number;
  pushSubscriptionsTotal: number;
  recentUsers: AdminRecentUser[];
}

export type PushSubscriptionKeys = { [key: string]: unknown };

export interface PushSubscription {
  id: string;
  userId: string;
  endpoint: string;
  keys?: PushSubscriptionKeys;
  createdAt: string;
}

export interface StatsOverview {
  activeChallenges: number;
  totalPlayers: number;
  matchesToday: number;
  completedMatches: number;
  openSlots: number;
}

export type MatchHistoryEntryOutcome =
  (typeof MatchHistoryEntryOutcome)[keyof typeof MatchHistoryEntryOutcome];

export const MatchHistoryEntryOutcome = {
  win: "win",
  loss: "loss",
  disputed: "disputed",
} as const;

export interface MatchHistoryEntry {
  challengeId: string;
  title: string;
  mode: string;
  outcome: MatchHistoryEntryOutcome;
  opponent?: string | null;
  playedAt: string;
}

export type ListChallengesParams = {
  mode?: ListChallengesMode;
  status?: ListChallengesStatus;
  hasSlots?: boolean;
};

export type ListChallengesMode =
  (typeof ListChallengesMode)[keyof typeof ListChallengesMode];

export const ListChallengesMode = {
  "1v1": "1v1",
  "2v2": "2v2",
  "4v4": "4v4",
} as const;

export type ListChallengesStatus =
  (typeof ListChallengesStatus)[keyof typeof ListChallengesStatus];

export const ListChallengesStatus = {
  open: "open",
  full: "full",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;