# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Challenge Arena API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /challenges:
    get:
      operationId: listChallenges
      tags: [challenges]
      summary: List active challenges
      parameters:
        - name: mode
          in: query
          schema:
            type: string
            enum: [1v1, 2v2, 4v4]
        - name: status
          in: query
          schema:
            type: string
            enum: [open, full, in_progress, completed, cancelled]
        - name: hasSlots
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of challenges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Challenge"

    post:
      operationId: createChallenge
      tags: [challenges]
      summary: Create a new challenge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChallengeBody"
      responses:
        "201":
          description: Challenge created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Challenge"
...
  /admin/overview:
    get:
      operationId: getAdminOverview
      tags: [admin]
      summary: Admin dashboard counts and recent signups
      responses:
        "200":
          description: Overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOverview"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: List all users with profile fields and stats (no passwords)
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/challenges:
    get:
      operationId: getAdminChallenges
      tags: [admin]
      summary: List challenges with full team/roster detail (latest first)
      responses:
        "200":
          description: Challenges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Challenge"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/notifications:
    get:
      operationId: getAdminNotifications
      tags: [admin]
      summary: Recent in-app notifications (all users)
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/match-results:
    get:
      operationId: getAdminMatchResults
      tags: [admin]
      summary: Recent match result submissions
      responses:
        "200":
          description: Match results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchResult"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/push-subscriptions:
    get:
      operationId: getAdminPushSubscriptions
      tags: [admin]
      summary: Web push subscription records
      responses:
        "200":
          description: Push subscriptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PushSubscription"
        "403":
          description: Not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /users/{userId}/stats:
    get:
      operationId: getUserStats
      tags: [stats]
      summary: Get player stats
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Player stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerStats"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    RegisterBody:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
        password:
          type: string
          minLength: 6
        email:
          type: string
          format: email
        freefireUid:
          type: string
        ign:
          type: string
        gender:
          type: string
          enum: [male, female, other]
          nullable: true
      required:
        - username
        - password
        - freefireUid
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        freefireUid:
          type: string
          nullable: true
        ign:
          type: string
          nullable: true
        gender:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        email:
          type: string
          nullable: true
        whatsappPhone:
          type: string
          nullable: true
        isAdmin:
          type: boolean
          description: True when user may access admin API (env or Firestore flag)
        stats:
          $ref: "#/components/schemas/PlayerStats"
          nullable: true
      required:
        - id
        - username
        - createdAt
...
    Challenge:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        mode:
          type: string
          enum: [1v1, 2v2, 4v4]
        scheduledAt:
          type: string
          format: date-time
        rules:
          type: array
          items:
            type: string
        customRule:
          type: string
          nullable: true
        status:
          type: string
          enum: [open, full, in_progress, completed, cancelled, disputed]
        teamA:
          $ref: "#/components/schemas/Team"
        teamB:
          $ref: "#/components/schemas/Team"
          nullable: true
        roomId:
          type: string
          nullable: true
        roomPassword:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        creatorId:
          type: string
        winnerId:
          type: string
          nullable: true
      required:
        - id
        - title
        - mode
        - scheduledAt
        - rules
        - status
        - teamA
        - createdAt
        - creatorId
    Team:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        leaderId:
          type: string
        players:
          type: array
          items:
            $ref: "#/components/schemas/TeamMember"
        maxSize:
          type: integer
      required:
        - id
        - name
        - leaderId
        - players
        - maxSize
    TeamMember:
      type: object
      properties:
        userId:
          type: string
        username:
          type: string
        ign:
          type: string
          nullable: true
        isLeader:
          type: boolean
        joinedAt:
          type: string
          format: date-time
      required:
        - userId
        - username
        - isLeader
        - joinedAt
    CreateChallengeBody:
      type: object
      properties:
        title:
          type: string
        mode:
          type: string
          enum: [1v1, 2v2, 4v4]
        scheduledAt:
          type: string
          format: date-time
        rules:
          type: array
          items:
            type: string
        customRule:
          type: string
          nullable: true
        teamName:
          type: string
      required:
        - title
        - mode
        - scheduledAt
        - rules
        - teamName
...
    MatchResult:
      type: object
      properties:
        id:
          type: string
        challengeId:
          type: string
        submittedBy:
          type: string
        winningSide:
          type: string
        screenshotUrl:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, confirmed, disputed]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - challengeId
        - submittedBy
        - winningSide
        - status
        - createdAt
...
    Notification:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
          enum: [player_joined, team_almost_full, match_ready, match_starting, match_result]
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        challengeId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - type
        - title
        - message
        - isRead
        - createdAt
...
    AdminRecentUser:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        ign:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - createdAt
    AdminOverview:
      type: object
      properties:
        usersTotal:
          type: integer
        challengesTotal:
          type: integer
        challengesByStatus:
          type: object
          additionalProperties:
            type: integer
        teamsTotal:
          type: integer
        teamMembersTotal:
          type: integer
        messagesTotal:
          type: integer
        notificationsTotal:
          type: integer
        unreadNotifications:
          type: integer
        matchResultsTotal:
          type: integer
        pushSubscriptionsTotal:
          type: integer
        recentUsers:
          type: array
          items:
            $ref: "#/components/schemas/AdminRecentUser"
      required:
        - usersTotal
        - challengesTotal
        - challengesByStatus
        - teamsTotal
        - teamMembersTotal
        - messagesTotal
        - notificationsTotal
        - unreadNotifications
        - matchResultsTotal
        - pushSubscriptionsTotal
        - recentUsers
    PushSubscription:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        endpoint:
          type: string
        keys:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - endpoint
        - createdAt
...
    PlayerStats:
      type: object
      properties:
        userId:
          type: string
        matchesPlayed:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        winStreak:
          type: integer
        weeklyWins:
          type: integer
      required:
        - userId
        - matchesPlayed
        - wins
        - losses
        - winStreak
        - weeklyWins
...